+++ snips_nlu/tests/test_config.py
@@ -112,8 +112,7 @@ class TestConfig(SnipsTest):
             "unit_name": LogRegIntentClassifier.unit_name,
             "data_augmentation_config":
                 IntentClassifierDataAugmentationConfig().to_dict(),
-            "featurizer_config": FeaturizerConfig().to_dict(),
-            "random_seed": 42
+            "featurizer_config": FeaturizerConfig().to_dict()
         }
 
         # When
@@ -151,8 +150,7 @@ class TestConfig(SnipsTest):
                 "algorithm": "lbfgs"
             },
             "data_augmentation_config":
-                SlotFillerDataAugmentationConfig().to_dict(),
-            "random_seed": 43
+                SlotFillerDataAugmentationConfig().to_dict()
         }
 
         # When